chore: Move `get_default_url` to `src/utils.rs`
This is slightly more elegant, even though this isn't being used anywhere other than the examples. It's a tradeoff that we are fine making.
