+++ examples/price_printer/main.rs
@@ -1,6 +1,5 @@
 extern crate tycho_simulation;
-#[path = "../shared/utils.rs"]
-mod shared_utils;
+
 mod ui;
 pub mod utils;
 
@@ -8,7 +7,6 @@ use std::{env, str::FromStr};
 
 use clap::Parser;
 use futures::{future::select_all, StreamExt};
-use shared_utils::get_default_url;
 use tokio::{sync::mpsc, task::JoinHandle};
 use tycho_client::feed::component_tracker::ComponentFilter;
 use tycho_common::models::Chain;
@@ -28,7 +26,7 @@ use tycho_simulation::{
         stream::ProtocolStreamBuilder,
     },
     protocol::models::Update,
-    utils::load_all_tokens,
+    utils::{get_default_url, load_all_tokens},
 };
 
 #[derive(Parser)]
