--- examples/shared/utils.rs
@@ -1,10 +1,0 @@
-use tycho_common::models::Chain;
-
-pub fn get_default_url(chain: &Chain) -> Option<String> {
-    match chain {
-        Chain::Ethereum => Some("tycho-beta.propellerheads.xyz".to_string()),
-        Chain::Base => Some("tycho-base-beta.propellerheads.xyz".to_string()),
-        Chain::Unichain => Some("tycho-unichain-beta.propellerheads.xyz".to_string()),
-        _ => None,
-    }
-}
